The base.html error resolved
Finally all errors have been successfully resolved and the website is functioning. The error emanated due to jinja2 templating language and the css being in templates folder. The css file should be in static folder(previously it was in templates folder which is what was causing error). Also when defining a block in html make sure the endblock column is not on the same line as block statement line.